requirements.txt
elasticsearch>=8.12
faker
python-dotenv
openai>=1.12
streamlit>=1.30
requests>=2.31
